serve pretty success and error pages on authentication callback
