+++ .gitignore
@@ -4,6 +4,7 @@
 .token
 .tmp*
 .venv/
+.hide/
