+++ local/static/templates/error.html
@@ -1,0 +1,61 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>404 Authentication Error</title>
+</head>
+<style>
+    /* TITLES */
+    h1{
+        color: #FFFFFF;
+        text-shadow: -0.05em 0.05em rgb(158, 168, 218), -0.1em 0.1em rgb(84, 96, 230), -0.15em 0.15em rgb(29, 43, 197), -0.2em 0.2em rgb(7, 16, 118);
+        font-size: 7em;
+        text-align: right;
+        font-style: italic;
+        font-family:Arial, Helvetica, sans-serif;
+        font-weight: bolder;
+    }
+    h2{
+        color: #FFFFFF;
+        text-align: left;
+        font-style: italic;
+        font-family:Arial, Helvetica, sans-serif;
+        font-weight: bolder;
+        text-shadow: -0.05em 0.05em rgb(84, 96, 230), -0.1em 0.1em rgb(7, 16, 118);
+        font-size: 5em;
+    }
+    /* TEXT */
+    p{
+        color: #FFFFFF;
+        font-size: 2em;
+        font-family:Arial, Helvetica, sans-serif;
+        font-weight: bolder;
+        text-align: center;
+        text-shadow: -0.05em 0.05em rgb(84, 96, 230), -0.1em 0.1em rgb(7, 16, 118);
+    }
+    body{
+        background-color: #0E1117;
+        background-image: url("https://perhalvorsen.com/static/img/background.jpg");
+    }
+    div.container{
+        background-color: rgba(63, 63, 63, 0.5);
+        border-radius: 10px;
+        padding: 20px;
+        margin: 20px;
+    }
+
+</style>
+<body>
+    <div class="container">
+        <h1>Uh oh, something went wrong.</h1>
+        <p>
+            Here's the error message from the back end: <code>{{ message }}</code>
+        </p>
+        <h2>
+            You can close this page and check the logs for more information.
+        </h2>
+    </div>
+
+</body>
+</html>
